feat: add all items to the item list and rewrote it in deathbulge-compliant format
