--- worlds/deathbulge/docs/setup_en.md
@@ -1,32 +1,0 @@
-# Your game here!
-
-## Required Software
-
-- A legally obtained copy of the game your are randomizing. Include the version if it needs to be specific (IE: NTSC/NA/Steam/Whatever else)
-- Anything else!
-  - You can use [this syntax to add hyperlinks to emulators or other web pages](https://www.ikea.com/us/en/p/blahaj-soft-toy-shark-90373590/)
-  - Be specific! Include versions! You'll still get asked why it's broken and it'll be a versioning issue but hey you can try
-- The built-in Archipelago client, which can be installed [here](https://github.com/ArchipelagoMW/Archipelago/releases)
-
-## Configuring your YAML file
-
-### What is a YAML file and why do I need one?
-
-You can leave this alone since it will be the same for every game so just delete this! I'm just here for explanations.
-
-Your YAML file contains a set of configuration options which provide the generator with information about how it should
-generate your game. Each player of a multiworld will provide their own YAML file. This setup allows each player to enjoy
-an experience customized for their taste, and different players in the same multiworld can all have different options.
-
-### Where do I get a YAML file?
-
-You can customize your options by visiting the [Your Game Here Options Page](/games/Your%20Game%20Here/player-options).
-
-Another hyperlink to the options page. The %20 in the link is a space. Make sure it matches the name you put for the your_game_here for the en_your_game_here.md doc.
-
-### Connect to the MultiServer
-
-1. Just describe how to get it connected
-
-2. Include any optional steps
-   - You can use this syntax to add sub steps
