--- worlds/deathbulge/options.py
@@ -1,36 +1,0 @@
-from dataclasses import dataclass
-from Options import Toggle, Range, Choice, PerGameCommonOptions
-
-
-class StartWithFastTravel(Toggle):
-    """Activates the fast travel earlier than in the base game. (Not Implemented)"""
-
-    display_name = "Start with fast Travel"  # this is the option name as it's displayed to the user on the webhost and in the spoiler log
-
-
-class TeamStart(Choice):
-    """Allows to start with more team members than vanilla. (Not Implemented)."""
-
-    display_name = "Team start"
-    option_faye = 0
-    option_ian = 1
-    option_briff = 2
-    alias_faye_and_ian = 1
-    alias_all = 2
-    default = 0
-
-
-class StartingMoney(Range):
-    """Set the starting money of your party. (Not Implemented)"""
-
-    display_name = "Stating money"
-    range_start = 0
-    range_end = 10000
-    default = 0
-
-
-@dataclass
-class DeathbulgeOptions(PerGameCommonOptions):
-    start_with_fast_travel: StartWithFastTravel
-    team_start: TeamStart
-    starting_money: StartingMoney
